requirements.txt
Flask==1.0.2
flask-restplus==0.12.1
Keras==2.2.4
tensorflow==1.11.0
Pillow==5.1.0
Werkzeug==0.16.0
